refactor: 전역에서 사용하는 util 패키지 이동
