+++ src/test/java/lotto/util/NumberFormatterTest.java
@@ -1,10 +1,12 @@
-package lotto.view.util;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import lotto.util.NumberFormatter;
 
 class NumberFormatterTest {
